requirements.txt
requests
python-dotenv
schedule
reportlab
matplotlib
opensearch-py
